refactor(config): unify ClusterConf into curvine-config (#1243)
Move full server ClusterConf/MasterConf/WorkerConf/JournalConf/DBConf
into curvine-config and leave curvine-common::conf as a re-export facade.
