+++ crates/common/curvine-config/Cargo.toml
@@ -14,11 +14,14 @@ curvine-core-error = { workspace = true }
 curvine-net = { workspace = true }
 curvine-rpc = { workspace = true }
 curvine-runtime = { workspace = true }
+curvine-fault = { workspace = true }
+curvine-io = { workspace = true }
 serde = { workspace = true }
 toml = { workspace = true }
 clap = { workspace = true }
 log = { workspace = true }
 nix = { workspace = true, features = ["net"] }
+regex = { workspace = true }
 
 [features]
 default = []
