+++ crates/common/curvine-config/src/master_conf.rs
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::conf::ClusterConf;
-use crate::rocksdb::DBConf;
-use orpc::common::{ByteUnit, DurationUnit, LogConf, Utils};
-use orpc::runtime::GroupExecutor;
-use orpc::{err_box, CommonResult};
+use crate::{ClusterConf, DBConf};
+use curvine_core_error::{err_box, CommonResult};
+use curvine_runtime::common::{ByteUnit, DurationUnit, LogConf, Utils};
+use curvine_runtime::runtime::GroupExecutor;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
