+++ crates/common/curvine-config/src/worker_conf.rs
@@ -14,11 +14,11 @@
 
 #![allow(clippy::should_implement_trait)]
 
-use crate::conf::ClusterConf;
 use crate::state::StorageType;
-use orpc::common::{ByteUnit, DurationUnit, FileUtils, LogConf, Utils};
-use orpc::io::SpdkConf;
-use orpc::{err_box, CommonResult};
+use crate::ClusterConf;
+use curvine_core_error::{err_box, CommonResult};
+use curvine_io::SpdkConf;
+use curvine_runtime::common::{ByteUnit, DurationUnit, FileUtils, LogConf, Utils};
 use regex::Regex;
 use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Default, Deserialize, PartialEq)]
